+++ example/test/variables/global.vars.ts
@@ -4,5 +4,9 @@ export default {
         firstName: 'James',
         lastName: 'Dean',
         password: 'itsASecretShh...'
-    }
+    },
+    email: 'james.dean@gmail.com',
+    firstName: 'James',
+    lastName: 'Dean',
+    password: 'itsASecretShh...'
 };