+++ example/test/variables/scenarioOutlineNested.vars.dev.ts
@@ -1,7 +1,7 @@
 export default {
     user: {
         email: 'james.dean@gmail.com',
-        firstName: 'James',
+        firstName: 'James|',
         lastName: 'Dean',
         password: 'itsASecretShh...'
     }