+++ example/test/variables/scenarioOutlineNested.vars.ts
@@ -1,0 +1,8 @@
+export default {
+    user: {
+        email: 'james.dean@gmail.com',
+        firstName: 'James|',
+        lastName: 'Dean',
+        password: 'itsASecretShh...'
+    }
+};